Added Wordle Solution UserScript
Installing this user script will allow you to hover over the Wordle logo at the top of the page and reveal the daily Wordle word to you. No longer will you have to sweat it out on the sixth move, praying to the gods of the lexicon to be on your side.